requirements.txt
flask==3.1.2
pandas
matplotlib
seaborn
nltk
scikit-learn
wordcloud
sumy
google-generativeai

